+++ samples/System.IO.Pipelines.Samples/HttpServer/HttpRequestStream.cs
@@ -101,7 +101,7 @@ namespace System.IO.Pipelines.Samples.Http
 
         private ValueTask<int> ReadAsync(ArraySegment<byte> buffer)
         {
-            return _connection.Input.ReadAsync(new Span<byte>(buffer.Array, buffer.Offset, buffer.Count));
+            return _connection.Input.ReadAsync(buffer);
         }
 
 #if NET451
