+++ src/System.IO.Pipelines.Extensions/PipelineReaderExtensions.cs
@@ -8,6 +8,90 @@ namespace System.IO.Pipelines
 {
     public static class PipelineReaderExtensions
     {
+        public static ValueTask<int> ReadAsync(this IPipeReader input, ArraySegment<byte> destination)
+        {
+            while (true)
+            {
+                var awaiter = input.ReadAsync();
+
+                if (!awaiter.IsCompleted)
+                {
+                    break;
+                }
+
+                var result = awaiter.GetResult();
+                var inputBuffer = result.Buffer;
+
+                var sliced = inputBuffer.Slice(0, Math.Min(inputBuffer.Length, destination.Count));
+                sliced.CopyTo(destination);
+
+                int actual = sliced.Length;
+                input.Advance(sliced.End);
+
+                if (actual != 0)
+                {
+                    return new ValueTask<int>(actual);
+                }
+
+                if (result.IsCompleted)
+                {
+                    return new ValueTask<int>(0);
+                }
+            }
+
+            return new ValueTask<int>(input.ReadAsyncAwaited(destination));
+        }
+
+        public static async Task CopyToAsync(this IPipeReader input, IPipeWriter output)
+        {
+            while (true)
+            {
+                var result = await input.ReadAsync();
+                var inputBuffer = result.Buffer;
+
+                try
+                {
+                    if (inputBuffer.IsEmpty && result.IsCompleted)
+                    {
+                        return;
+                    }
+
+                    var buffer = output.Alloc();
+
+                    buffer.Append(inputBuffer);
+
+                    await buffer.FlushAsync();
+                }
+                finally
+                {
+                    input.Advance(inputBuffer.End);
+                }
+            }
+        }
+
+        private static async Task<int> ReadAsyncAwaited(this IPipeReader input, ArraySegment<byte> destination)
+        {
+            while (true)
+            {
+                var result = await input.ReadAsync();
+                var inputBuffer = result.Buffer;
+
+                var sliced = inputBuffer.Slice(0, Math.Min(inputBuffer.Length, destination.Count));
+                sliced.CopyTo(destination);
+                int actual = sliced.Length;
+                input.Advance(sliced.End);
+
+                if (actual != 0)
+                {
+                    return actual;
+                }
+                else if (result.IsCompleted)
+                {
+                    return 0;
+                }
+            }
+        }
+
         public static Task CopyToAsync(this IPipeReader input, Stream stream)
         {
             return input.CopyToAsync(stream, 4096, CancellationToken.None);
